Revert "Merge branch 'main' into dev"
This reverts commit 19dcafe3fb1af4d186eb49a9018582c0011080d4, reversing
changes made to 102fe5e6a77c9077c735b8fedaf44e86520b9337.
