+++ frontend/src/lib/api/auth.ts
@@ -1,7 +1,7 @@
+// Замените хардкод URL на переменную окружения
 const API_URL = import.meta.env.PROD 
   ? '/api'  // В production используем относительный путь
   : import.meta.env.VITE_API_URL || 'http://localhost:8090';
-
 
 export async function sendEmailCode(email: string): Promise<boolean> {
   try {
@@ -69,30 +69,5 @@ export async function login(email: string, password: string) {
   } catch (error) {
     console.error('Ошибка входа:', error);
     return false;
-  }
-}
-
-
-
-
-export async function fetchUserProfile() {
-  const token = sessionStorage.getItem('accessToken');
-  if (!token) return null;
-
-  try {
-    const response = await fetch('/api/users/me', {
-      headers: {
-        'Authorization': `Bearer ${token}`
-      }
-    });
-    if (response.ok) {
-      const userData = await response.json();
-      currentUser.set(userData);
-      return userData;
-    }
-    return null;
-  } catch (error) {
-    console.error('Failed to fetch user profile:', error);
-    return null;
   }
 }
